+++ crates/sdmx_json/src/macros/item_scheme.rs
@@ -1,16 +1,18 @@
 macro_rules! impl_item_scheme {
-	($(($struct: ty, $field: ident)),+ $(,)?) => {
+	($(($struct: ty, $field: ident, $item: ty)),+ $(,)?) => {
 		$(
 			impl $crate::structure::ItemScheme for $struct {
+				type Item = $item;
+
 				fn is_partial(&self) -> Option<bool> {
 					self.is_partial
 				}
 
-				fn items(&self) -> Option<&Vec<$crate::structure::Item>> {
+				fn items(&self) -> Option<&Vec<Self::Item>> {
 					self.$field.as_ref()
 				}
 
-				fn set_items(&mut self, items: Option<Vec<$crate::structure::Item>>) {
+				fn set_items(&mut self, items: Option<Vec<Self::Item>>) {
 					self.$field = items;
 				}